refactor Event
